optimize src

+++ NetExercise.Test/Basic/Exercise08Test.cs
@@ -5,8 +5,8 @@ namespace NetExercise.Test.Basic
 {
     public class Exercise08Test
     {
-        [TestCase(3.1415f)]
-        public void Test_Pi_Ok(float result)
+        [TestCase(3.1415)]
+        public void Test_Pi_Ok(double result)
         {
             Assert.AreEqual(Exercise08.Pi(), result,0.0001);
         }
